refactor: projects -> dashboard
To avoid overlap of terms (project v checker), rename the projects
page and supporting cast to dashboard
