+++ src/client/pages/index.ts
@@ -2,5 +2,5 @@ export { Landing } from './Landing'
 export { Checker } from './Checker'
 export { Debug } from './Debug'
 export { Login } from './Login'
-export { Projects } from './Projects'
+export { Dashboard } from './Dashboard'
 export { FormBuilder } from './FormBuilder'
